Moving plugin models to their own module
